+++ entityparsing.cpp
@@ -38,13 +38,13 @@ namespace EntityParsing {
         int x1col = -1, y1col = -1, x2col = -1, y2col = -1;
         for (i = 0; i < strings.size(); i++) {
             if (strings[i] == "x1") {
-                x1col = i;
+                x1col = static_cast<int>(i);
             } else if (strings[i] == "x2") {
-                x2col = i;
+                x2col = static_cast<int>(i);
             } else if (strings[i] == "y1") {
-                y1col = i;
+                y1col = static_cast<int>(i);
             } else if (strings[i] == "y2") {
-                y2col = i;
+                y2col = static_cast<int>(i);
             }
         }
 
@@ -108,9 +108,9 @@ namespace EntityParsing {
         int xcol = -1, ycol = -1;
         for (i = 0; i < strings.size(); i++) {
             if (strings[i] == "x") {
-                xcol = i;
+                xcol = static_cast<int>(i);
             } else if (strings[i] == "y") {
-                ycol = i;
+                ycol = static_cast<int>(i);
             }
         }
 
@@ -180,13 +180,13 @@ namespace EntityParsing {
         int xcol = -1, ycol = -1, anglecol = -1, viewcol = -1;
         for (i = 0; i < strings.size(); i++) {
             if (strings[i] == "x") {
-                xcol = i;
+                xcol = static_cast<int>(i);
             } else if (strings[i] == "y") {
-                ycol = i;
+                ycol = static_cast<int>(i);
             } else if (strings[i] == "angle") {
-                anglecol = i;
+                anglecol = static_cast<int>(i);
             } else if (strings[i] == "viewangle") {
-                viewcol = i;
+                viewcol = static_cast<int>(i);
             }
         }
 
@@ -210,12 +210,14 @@ namespace EntityParsing {
                     throw EntityParseException(message.str().c_str());
                 }
 
-                double x = std::atof(strings[xcol].c_str());
-                double y = std::atof(strings[ycol].c_str());
+                double x = std::atof(strings[static_cast<size_t>(xcol)].c_str());
+                double y = std::atof(strings[static_cast<size_t>(ycol)].c_str());
 
                 if (partialIsovists) {
-                    double angle = std::atof(strings[anglecol].c_str()) / 180.0 * M_PI;
-                    double viewAngle = std::atof(strings[viewcol].c_str()) / 180.0 * M_PI;
+                    double angle =
+                        std::atof(strings[static_cast<size_t>(anglecol)].c_str()) / 180.0 * M_PI;
+                    double viewAngle =
+                        std::atof(strings[static_cast<size_t>(viewcol)].c_str()) / 180.0 * M_PI;
                     isovists.push_back(IsovistDefinition(x, y, angle, viewAngle));
                 } else {
                     isovists.push_back(IsovistDefinition(x, y));
